🔄 Use template instead of clone
Refactor the package contents to rely on a `copier` template instead of cloning
the repository. This will make it easier for workshop participants to get started
with their own package. Some of the more advanced topics, such as CI, are also
easier to execute in case they are working with their own package.